requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.21.0
opencv-python>=4.6.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
pillow>=9.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.5.0
pandas>=1.4.0
scipy>=1.9.0
tensorflow-gpu>=2.10.0
jupyter>=1.0.0
